chatbot-system/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# LLM and AI Orchestration
langchain==0.1.0
langchain-core>=0.1.7,<0.2
langchain-community>=0.0.28  # For SQL Agent toolkit
langchain-anthropic>=0.3.21  # Anthropic provider for LangChain
langgraph>=0.0.35  # Improved StateGraph and multi-agent support
pydantic>=2.10.0
pydantic-settings>=2.1.0
numpy==1.26.3
scikit-learn==1.4.0

# Database Connectors
sqlalchemy>=2.0.36
oracledb==2.0.0
asyncpg>=0.30.0
redis==5.0.1

# HTTP and Async
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3
zeep==4.2.1
aiohttp>=3.9.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
watchdog==3.0.0

# Development Tools
black==23.12.1
pylint==3.0.3
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
websocket-client==1.7.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-json-logger==2.0.7
tqdm>=4.67.0  # Progress bar for bulk testing

# LLM Provider SDKs
anthropic>=0.69.0  # Official Anthropic SDK (Claude)
# openai==1.10.0  # Official OpenAI SDK (uncomment if needed)
# litellm==1.17.0  # Unified LLM interface (100+ providers)

# Optional Observability (for production)
# sentry-sdk[fastapi]==1.40.0  # Error tracking
# prometheus-client==0.19.0  # Metrics
# opentelemetry-api==1.22.0  # Distributed tracing
# opentelemetry-sdk==1.22.0
# opentelemetry-instrumentation-fastapi==0.43b0